+++ packages/sdk/src/logs/index.ts
@@ -3,5 +3,6 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+export type { Instrumentation } from '@opentelemetry/instrumentation';
 export type { LogsConfig, WebSdk } from '../core/types.ts';
 export { startLogsSdk } from './startLogsSdk.ts';