+++ packages/sdk/src/traces/startTracesSdk.ts
@@ -6,13 +6,16 @@
 import { context, diag, propagation, trace } from '@opentelemetry/api';
 import { CompositePropagator } from '@opentelemetry/core';
 import { OTLPTraceExporter } from '@opentelemetry/exporter-trace-otlp-http';
+import { registerInstrumentations } from '@opentelemetry/instrumentation';
 import {
   defaultResource,
   resourceFromAttributes,
 } from '@opentelemetry/resources';
 import type { SpanProcessor } from '@opentelemetry/sdk-trace';
 import { BatchSpanProcessor, TracerProvider } from '@opentelemetry/sdk-trace';
+import { getDefaultContextManager } from '../core/context.ts';
 import { setSdkLogger } from '../core/diag.ts';
+import { getDefaultPropagators } from '../core/propagation.ts';
 import type { TracesConfig, WebSdk } from '../core/types.ts';
 
 const DEFAULT_TRACES_OTLP_ENDPOINT = 'http://localhost:4318/v1/traces';
@@ -75,26 +78,31 @@ export function startTracesSdk(config?: TracesConfig): WebSdk {
     return NOOP_SDK;
   }
   const tracerProvider = new TracerProvider({
-    // sampler: new TraceIdRatioBasedSampler(
-    //   typeof config?.sampleRate === "number" ? config?.sampleRate : 1,
-    // ),
     resource,
     spanLimits: config?.spanLimits,
     spanProcessors,
+    sampler: config?.sampler,
   });
   trace.setGlobalTracerProvider(tracerProvider);
 
-  if (config?.propagators) {
-    const { propagators } = config;
-    propagation.setGlobalPropagator(new CompositePropagator({ propagators }));
-  }
+  const propagators = config?.propagators ?? getDefaultPropagators();
+  propagation.setGlobalPropagator(new CompositePropagator({ propagators }));
 
-  if (config?.contextManager) {
-    context.setGlobalContextManager(config.contextManager);
+  const contextManager = config?.contextManager ?? getDefaultContextManager();
+  contextManager.enable();
+  context.setGlobalContextManager(contextManager);
+
+  // Register instrumentations
+  let deregisterInstrumentations: (() => void) | undefined;
+  if (config?.instrumentations?.length) {
+    deregisterInstrumentations = registerInstrumentations({
+      instrumentations: config.instrumentations,
+    });
   }
 
   return {
     shutdown() {
+      deregisterInstrumentations?.();
       return tracerProvider.shutdown();
     },
   };