+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,39 @@
 # Changelog
 
+## [0.4.0](https://github.com/open-telemetry/opentelemetry-browser/compare/browser-sdk-v0.3.0...browser-sdk-v0.4.0) (2026-09-09)
+
+
+### Features
+
+* **sdk:** populate browser.document.url.full on all spans and log records ([#421](https://github.com/open-telemetry/opentelemetry-browser/issues/421)) ([a2b4cf8](https://github.com/open-telemetry/opentelemetry-browser/commit/a2b4cf8305eec57422045bd0b47aa42eafacffd8))
+
+
+### Bug Fixes
+
+* **deps:** update opentelemetry ([#419](https://github.com/open-telemetry/opentelemetry-browser/issues/419)) ([d1f765d](https://github.com/open-telemetry/opentelemetry-browser/commit/d1f765d08bbcdd120396ef8a82b3864ec92c82cc))
+
+## [0.3.0](https://github.com/open-telemetry/opentelemetry-browser/compare/browser-sdk-v0.2.0...browser-sdk-v0.3.0) (2026-08-27)
+
+
+### Features
+
+* **browser-sdk:** wire default context manager and propagators into startTracesSdk ([#393](https://github.com/open-telemetry/opentelemetry-browser/issues/393)) ([063dddd](https://github.com/open-telemetry/opentelemetry-browser/commit/063dddd00fdc64cdfbe0532b55b10278341710d0))
+
+## [0.2.0](https://github.com/open-telemetry/opentelemetry-browser/compare/browser-sdk-v0.1.0...browser-sdk-v0.2.0) (2026-07-29)
+
+
+### Features
+
+* add e2e test infrastructure with ErrorsInstrumentation example ([#332](https://github.com/open-telemetry/opentelemetry-browser/issues/332)) ([47bf5d6](https://github.com/open-telemetry/opentelemetry-browser/commit/47bf5d6e0379914b1fb2e14dc23cb5a293009759))
+
+
+### Bug Fixes
+
+* **deps:** update opentelemetry ([#363](https://github.com/open-telemetry/opentelemetry-browser/issues/363)) ([533fcdf](https://github.com/open-telemetry/opentelemetry-browser/commit/533fcdf72db5eed6158b9e2bc42deb2452236d98))
+* **sdk:** accept sampler option in traces SDK ([#369](https://github.com/open-telemetry/opentelemetry-browser/issues/369)) ([c729d8c](https://github.com/open-telemetry/opentelemetry-browser/commit/c729d8c11c9bf210981be80df7591fa8521ebee4))
+
+## Changelog
+
 ### Features
 
 ### Bug Fixes